refactor: move personal controller to separate folder
