+++ src/controllers/Controller.ts
@@ -24,11 +24,7 @@ export default class Controller implements IController {
 
   public static getInstance = (): IController => this.instance;
 
-  public execute = (
-    data: RawData,
-    socketMap: Record<string, WebSocket>,
-    socketId: string
-  ): void => {
+  public execute = (data: RawData, socket: WebSocket): void => {
     const parsedData: IData = parseRawData(data);
     const controllerName: string | undefined = this.findCommand(
       parsedData.type
@@ -38,14 +34,9 @@ export default class Controller implements IController {
       if (!controllerName)
         throw new Error(`Command type '${parsedData.type}' doesn't exists!`);
 
-      this.commands[controllerName as ERespType].execute(
-        parsedData,
-        socketMap,
-        socketId
-      );
+      this.commands[controllerName as ERespType].execute(parsedData, socket);
     } catch (err) {
       const msg = extractErrorMsg(err, 'unknown error');
-      const socket: WebSocket | undefined = socketMap[socketId];
 
       if (!socket) return;
       socket.send(JSON.stringify({ error: true, errorText: msg }));
