+++ src/types/types.ts
@@ -1,5 +1,9 @@
-import { IPlayer } from './interfaces/IPlayer';
+import WebSocket from 'ws';
+
+import IData from '@src/types/interfaces/IData';
+import { IPlayer } from '@src/types/interfaces/IPlayer';
 
 type IPlayerParams = Pick<IPlayer, 'login' | 'password' | 'socket'>;
+type TypedCommand = (data: IData, socket: WebSocket) => void;
 
-export { IPlayerParams };
+export { IPlayerParams, TypedCommand };
